Added Homework 6 Folder
Added Homework 6 Folder
Added HTML files for 3 pages of my website. I'm not sure if the hyperlinks will work from this link. But all 3 files are there and worked when I coded them. I'm just not sure if they'll work because they have been moved and the files were locally on my computer.